Upgrade to Kubernetes 1.10.0 libraries (#477)
* Update all Kubernetes deps to 1.10.0-rc.1

To ensure all k8s dependencies are locked to the same point, each one is
now constrained to the same tag generated from a single tagged
commit in the kubernetes/kubernetes repo. Updating client-go to
release-7.0 causes issues because it doesn't correspond to a tag in any
of the other repos (it diverges more than 100 commits from the
1.10.0-rc.1 tag).

The k8s.io/gengo dependency doesn't use the same tag as the others, so
it's updated to its current master.

1.10.0-rc.1 is the most recent tag in the k8s repos. When 1.10 is
released, we should change the tag to pull in any changes since rc1.

* Regen code with Kubernetes 1.10

* Use RaceFreeFakeWatcher in ObjectTracker

The FakeWatcher allows sends on the result channel after it's closed,
for example calling Stop() then Add() will panic. RaceFreeFakeWatcher
checks whether the watcher is stopped before attempting to send. It also
panics instead of blocking when the result channel is full.

This commit is copied from kubernetes/kubernetes#61195 where it is
currently unmerged. It's necessary to avoid our tests panicking due
to sends on closed channels.

* Patch Kubernetes client when updating deps

This patch applied here is the response body of
https://github.com/kubernetes/kubernetes/pull/61195.patch, and also the
same patch as applied in the parent commit. The patch is necessary to avoid
panics when watches fire in Kubernetes fake clients.

Dep will revert the patch every time it runs, so we need to re-patch in
hack/update-deps.sh immediately after dep finishes.

* Update to kubernetes-1.10.0 release tag